+++ android/src/main/kotlin/io/flutter/embedding/android/ThrioFlutterActivityBase.kt
@@ -35,6 +35,8 @@ interface ThrioFlutterActivityBase {
 
     fun cleanUpFlutterEngine(flutterEngine: FlutterEngine)
 
+    fun onBackPressed()
+
     fun shouldMoveToBack(): Boolean = true
 
     fun shouldDestroyEngineWithHost(): Boolean