+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - feat: android support fragment
 - fix: routeSettings for NavigatorPageView not refresh
 - feat: make page view lifecycle work for NavigatorPageObserver 
+- feat: support root page WillPopScope
 
 ## 3.12.5